Parse query elements out of search-string as needed
Under react-router v4.0.0-alpha.6, which is what we were using until
recently, the location object has parsed-out query parameters in the
`query` element. In v4.0 proper, that is no longer provided. So in the
absence of a pre-parsed query, we now do it ourselves, using
queryString.parse(location.search).

Fixes STRIPES-258.