First listed settings sub-section is shown
When no sub-section has yet been chosen.

This is a fix for STRIPES-357.

But I am not happy with it yet, because the selected (first)
sub-section listed in the section column is not highlighted. I am not
sure how to make that happen, but I am committing what I have so far
since it's at least a step forward.