Lint: separate component, element props. See STCOM-21
In some cases props passed to a component are then pushed down to a
child element and we would have to filter out irrelevant props with code
like this:

const { junk1, junk2, junk3, ...inputProps } = this.props;
<input type="checkbox" {...inputProps} />

The junkX variables are created solely for the purpose of pulling those
keys out of inputProps, which causes ESLint to complain about unused
variables.

The function separateComponentProps provides a better way of handling
this situation without creating unused variables.