show overflow in wrapper elements. Refs UIU-128.
The wrapper elements around a list, in addition to the individual cells,
need to allow visible overflow in order for the cells' overflow settings
to bubble up. This is necessary to allow overflow in the trailing cells
of a list; otherwise their overflow is simply cropped by the wrapper
element.