restore "id" to the generated component
use of separateComponentProps() helps to clean up code by not littering
it with a bunch of local variables that exist for no other reason than
to pull out a key that would otherwise be left in a destructured object.
But it does have the side effect of not passing along key-value pairs to
the nested component if that key is defined on the parent component.
"id" is such a key here. So this commit fixes that, restoring the "id"
attribute on the generated component.