Get main params from anointed resource, not URL
qindex, filters, query, sort are now all drawn from the anointed
resource. When only a single application is running, this makes no
difference, but when running app-in-app (for example, a second
instance of Users to choose a proxy or sponsor), it ensures that the
correct instance's anointed resource is used.

Fixes STCOM-226.