lint
Just for the record, let it be known: two-character indentation is a
blight upon the landscape, a scourge upon the earth, an abomination. It
is so, so easy not to notice that things are off when you only have
two characters to work with. This has nothing to do with any particular
style guide; it's just too hard to follow the flow through nested conditional
expressions when only two characters set them off from each other. If I
had to choose a single stupid convention to change, this would be it.

STUPID STUPID STUPID STUPID STUPID STUPID STUPID STUPID STUPID STUPID