use code -> name mapping in AddressFieldGroup
The previous implementation of AddressFieldGroup mapped country-name to
country-name in its select-list of countries. This required every
component that used this one to translate from country-name to
country-code in order to properly store the country-code on the
back-end, and it broke down in the case of translated country names.

Fixes STCOM-242.