use column-header values, not keys, for aria labels
Now that we are implementing i18n strings consistently, we are typically
passing in a list map of field-names to column-labels, rather than just
using the column-labels. The field-names are not translated and intended
to be machine-readable, e.g. patronGroup, not human friendly, e.g.
Patron Group. This change uses the column-labels.

Fixes STCOM-246