add hidden filter config. Refs UIU-400
Allow filters to have hidden configuration values. For example, suppose
a filter group "species" existed and the query should *always* include
humans and allow the user to choose whether Ewok, Jawa, or Wookie
members should also be included in the search. Ewok, Jawa, and Wookie
would be included in the "values" array of the config element with
name: "species" and Human would be included in the array of the
hiddenConfig element keyed by "species".