leave `activeLink` undefined when not provided
Providing a default value for `activeLink` when not provided via props
would cause problems on settings page with multiple sections and the
selected link outside the first section. If the active link is not in
the first section, the default value would be applied in section-one
(highlighting the first item) and then, because the sections are
rendered iteratively, `activeLink` could be assigned additional times in
subsequent sections.

Fixes UIORG-79