do not check component types
A PropTypes check for a particular component type may fail if that
component is, say, a functional component or one that returns a ref.
`PropTypes.node` allows us to detect that we're getting a renderable
thing without thinking too hard about it.

Why am I doing this? I want a clean console, and you should too.