Fix documentation of <Timepicker>'s timeZone prop
This was claimed to default to 'UTC' but actually the prop itself has
no default value, and the behaviour of the Timepicker defaults to
functioning within the timezone of the prevailing locale
(`props.intl.timeZone`).

Added a note to say that passing `timeZone='UTC'` allows you to treat
the time as an absolute time of day.