STCOM-659 provide FormattedDate, FormattedTime
ISO-8601 is ambiguous about how timezones in timestamps should be
formatted. In short, the spec says a timezone can be formatted as an
offset like `+HHMM` or `+HH:MM`, but the W3C opted to only require
browsers to implement the latter. Unfortunately, our backend APIs chose
to implement the former.

Some browsers handle both, and for them we are grateful. Others, like
Safari, or Chrome on iOS, do not, and for them we provide
`<FormattedDate>` and `<FormattedTime>` to provide that missing colon.

So much work for two little dots, eh? Sheesh.

Refs STCOM-659