STCOM-492: Changed FilterGroups operator from = to ==
Change the default `<FilterGroups>` operator from `=` to `==`. 

The `=` syntax is translated to `~ regex` search on database side and is 
very slow. Since filters search the exact key, we should use `==` syntax 
that translates to `LIKE` (equivalent to exact in most cases) and that is 
way faster. For example in location filtered instance search, we see 
performance improved about *30x*.

Refs STCOM-492