STCOM-817 sync timezones in timepicker tests (#1526)
Sync the timezone given to the `<Timepicker>` and used to check its
output. The dates when DST is observed are not consistent across
timezones, e.g. the US starts DST a week before and ends it a week after
most other countries in the world. Thus, this test would fail if it ran
during one of those weeks in an environment where the local timezone is
a non-US timezone.

Refs STCOM-817