STCOM-860 always use Arabic numerals given backendDateStandard
The implied promise of `outputFormatter`, given `backendDateStandard` is
that it will format dates using Arabic numerals (0-9) as this is what
the backend expects. Previously, it handled only for formatting but not
the conversion to Arabic ("Latn") numerals if the current locale used
non-Arabic numerals, which is the case for manu Arabic locales that use
Arabic-Hindi numerals.

Now, `<Datepicker>` delivers on both its promises:

1. it formats the date as expected for the backend
2. it provieds dates in Arabic numerals as expected for the backend

Refs STCOM-860, FOLIO-3208