bump version to match latest release
* `stripes-components` `9.2.1` was released on a branch while the version on master remained at `9.2.0`.
* `stripes` `6.2.2` was release _on master_, referencing `9.2.1`.
This will cause build problems if `stripes-components` is cloned into a workspace: the `stripes` dep on `~9.2.1` means the git-clone of `stripes-components`, still at `9.2.0`, doesn't satisfy the dependency and it will pull down an additional copy into its `node_modules`. This means there are now multiple copies of `stripes-components` in the bundle, and Bad Things happen when that occurs.