correctly test RadioButton focus under React 17 (#1641)
Focus handling changed under react 17. This was patched on master 
in #1630 in a commit that introduced a dependency on some additional 
@folio/stripes-testing components that have not yet been released. 
Thus, when that commit was picked to this release branch in e34c1bc, 
the release process failed. 

So: this PR reverts the portion of that commit that introduced new 
dependencies and keeps others parts that fixed the tests. Whew.