STCOM-902 interactors should not use dynamic css vars (#1668)
Interactors like
```
.${css.mclHeader}
```
may result in BTOG errors like
```
SyntaxError: ".mclHeader---IgPC+" is not a valid selector
```
because of the `+` appended to the end. I'm not sure exactly where it
comes from, but it ain't right since `+` is meaningful in
CSS-selector-land (as the error tells us). In any case, changing the
selectors to something like
```
[class*=mclHeader---]
```
avoids the problem.

Refs STCOM-902