STCOM-914 REALLY correct LayoutHeader documentation
The prop-types and the code didn't agree, so while the documentation was
changed in #1697 it actually made things worse because following it (as
in https://github.com/folio-org/stripes-smart-components/pull/1164)
would not work.

It appears that `AddressView` is the only case where `actions` is passed
to `LayoutHeader`, but leaving the existing implementation as-is and
changing the doc to match still feels like the safest bet.

Refs STCOM-914