STCOM-1056 avoid Paneset NPE when evaluating cached layouts
In production, we saw that `this.state.layoutCache.findIndex(...)` was
sometimes throwing an NPE on `layoutCache`. It isn't totally clear how
this happens since it's initialized from a prop with valid default
value. Maybe sometimes the prop is present but `null`? Dunno, but in any
case the `...findIndex()` code is now more cautious.

Aside: renamed `hasCachedLayout` to `cachedLayoutIndex` since the
previoius name gives the impression it returns a boolean when what it
actually returns is an array index, i.e. an int >= -1.

Refs STCOM-1056