STCOM-1072 upgrade postcss-* stack to latest available (#1938)
This update bumps postcss-* stack to latest versions.

Many of these confident dependencies have node: ">=10" as their engine dependency.
We've waited a while with a locked-in custom-properties dependency. They've added a flag to suppress the noisy error message. I've decided to apply it (in stripes-webpack) so that we can stay current on the version, and I feel like before they move anywhere, its possible that we could eliminate use of that plugin entirely (next spike-able effort)

Mergeable along with folio-org/stripes-webpack#86

Refs STCOM-1072