STCOM-1153 omit webpack dependency (#2044)
`webpack` is not used directly here and is supplied indirectly via
`@folio/stripes-cli`. Thus, rather than unpinning it, in fact it can be
completely removed. This will result in some dev-deps whining that their
peer-dep on `webpack` is no unsatisfied, but, again, it _is_ satisfied
by `@folio/stripes-cli`; we just don't have a way to indicate that.

Refs STCOM-1153