Bump @svgr/webpack from 7.0.0 to 8.1.0
This appears to be an update in name only. From what I can tell by
looking at the release notes, the 7 to 8 bump is a side-effect of
changes elsewhere in its monorepo, i.e. there are no actual changes
here, but it would be nice to keep the version current in case of future
development on the 8.x line.

This work must be done after folio-org/stripes-webpack/pull/122.