STCOM-1209 correctly handle multiple callouts
Correctly handle callout-removal when multiple callouts are added in
quick succession.

Previously, callouts were stored in an array in state and removal was
handled by iterating through the list to find an item given its id and
then removed via Array.splice on a cloned copy of state. When multiple
callouts were added in quick succession, this could lead to the wrong
callout being removed because of changes in the underlying state array.

Given callouts have a unique `id` property, storing them by `id` in an
object/set, instead of on a list, allows them to be correctly and
consistently manipulated.

h/t usavkov-epam for precisely describing the problem, and proposing a
solution, in UIOR-1145.

Refs STCOM-1209