fix: Change output of getBackendDateStandard when outputBackendValue is false
Previously getBackendDateStandard would return `undefined` when outputBackendValue was set to `fase. This value would then attempt to spread further down the code, causing a TypeError.

Instead replace this with an empty array []. This assumes that getBackendDateStandard output was not tested for being undefined anywhere, which appears to be the case, and so this can safely be changed.

STCOM-1372