STCOM-1379 pin currency-codes to avoid unexpected dupes
Data that was uniquely keyed up through `v2.1.0` is not so in `v2.2.0`
(there are multiple entries for the currency `Bolívar Soberano`). It is
not clear if the uniqueness property was intentional or incidental.

Since we have tests that directly rely on the data we import, we should
_always_ depend on a specific version in order to avoid this kind of
problem, and so we can choose when to address it instead of being forced
to address it whenever the third-party dependency happens to publish a
new release.

Refs STCOM-1379