STCOM-1391 Tooltip - check for hover on tooltip content before closing the tooltip. (#2411)
According to
[WCAG](https://www.w3.org/TR/WCAG21/#content-on-hover-or-focus), content
that's available/visible on hover should remain visible when the content
itself is hovered... mouse-readers everywhere unite!

### Approach
Tooltip already uses timeouts to provide a brief gap between event
listeners and state updates. This implementation added a timeout to
hiding the tooltip where we only hide the tooltip if neither the trigger
nor the tooltip element itself are hovered by the mouse (using
`matches(':hover')`). The CSS of the tooltip class had to be affected to
observe mouse events since they are all placed in the
`#OverlayContainer` by default.