STCOM-1336 remove react-overlays dependency (#2429)
## [STCOM-1336](https://folio-org.atlassian.net/browse/STCOM-1336)

This implements
[STCOM-1339](https://folio-org.atlassian.net/browse/STCOM-1339), as it
was a sub-task of this story.

`react-overlays` hasn't been published in 2 years.

After the modal refactor, the remaining bits were the 
`useRootClose` hook and an instance of `<Portal>`

- <Portal> was a simple refactor - `React-DOM`'s base functionality for
portals is all that's necessary there.

- ~~I opted to pull in a new dependency - `usehooks-ts` instead of
re-implementing `useRootClose` locally.
[usehooks-ts](https://usehooks-ts.com/introduction) is a minimal,
no-dependency lib with active development and a nice dev experience
curtesy of TS! It is pre-transpiled to commonjs, of course...
`useClickOutside` nicely replaced `onRootClose`.~~
- our own `useClickOutside` is implemented a bit differently, but served
purpose just fine rather than taking on any dependency - and the logic
is quite similar to that of the compound hook within `usehooks-ts`.


Additionally - updated shared workflows to 1.9 so that tests can run!