REPAIR AN EXTREMELY IMPORTANT MISCARRIAGE OF SPELLIG
This is the most important PR that has ever been issued. Go to Defcon-1.
Er, Defcon-5! Whatever the most important Defcon is, go to that one! Do
not pass go. Do not collect $200. Even if you had $200, it wouldn't be
enough to get you out of jail if the Self-Appointed Grammar Police found
out about this. Never you mind that this is about spelling, not grammer.
Argh! I mean `ar`! Grammar! Spelling, spelling, I mean spelling!

Q: What is a pirate's favorite letter?
A?: R!
A!: Everyone thinks it's R but it's the C!

Anyway, the Cs that were missing have been found.