STCOM-1426 make AutoSuggest::defaultIncludeItem() null-safe
`defaultIncludeItem(haystack, needle)` assumes `haystack` is shaped like
`{ value }` and returns true if the `value` attribute is a string
containing `needle`.

The previous incarnation did not validate whether `item` or `item.value`
were null, leading to NPEs if haystack was `null` or if it did not
contain an attribute named `value`.

Refs STCOM-1426