STCOM-1429 Unable to close popover (#2462)
When clicking an open Popover's target, the overlay will close and then
re-open. This was caused by the `<RootCloseWrapper`'s handler triggering
(closing the overlay) followed by the actual click handler (toggling the
open status closed -> open).

Approach:
Check for the target at the `<RootCloseWrapper>`'s handler and allow
Toggle to 'toggle' as it should.

Additional: Focus styles for the overlay since there wasn't any
indication, and adding `toggle` functionality to the overlay's escape
key handler so that it actually closes the dialog (otherwise it was
empty, aside from stopping propagation.)

Comments added for future us.