STCOM-1460 s/rehooks-ts/usehooks-ts/g
If we're lucky, updates #2495; replaces #2497

`rehooks-ts` blows up in CI in other repos when their tests that
leverage stripes-components claim not to be able to find `rehooks-ts`.
Maybe this is a jest hiccup tripped by the fact that `rehooks-ts` claims
to export `dist/index.cjs`, a file that does not exist. So that's ...
odd.

`usehooks-ts` appears to offer a hook that offers the same API. So
that's ... convenient!

Refs STCOM-1460