STCOM-814 Leverage sourcemaps in ErrorBoundary (#2496)
Summary

If a sourcemap can be loaded and parsed, show the mapped stacktrace
instead of the original (pretty much useless) stacktrace.

Details

* March 2021: We merged #1512 even though the reporting wasn't great 😑 
* ... ~200 "Something went wrong" tickets are filed 😱 
* October 2024: Ugh, STCOM-1355, we should purge the useless reporting 😢
* ... ~70 more "Something went wrong" tickets are filed 🙀 
* September 2025, WOLFCon's "How do we figure out what the problem
is?" reminds me how annoying it is that we couldn't do better, which led to
more searching, which led to
https://github.com/novocaine/sourcemapped-stacktrace, which has been
around since 2018 so I don't know why we didn't find it earlier, but we
didn't, and then we did, and here we are. So even though STCOM-814 was
closed as "Won't do" years ago, sometimes ya just gotta do it
anyway 🔥 .

Refs STCOM-814, STCOM-1355