STCOM-1436 - add OverlayWrapper to Accordionset (#2500)
Sure fire for when the overlays are overlapped by other stuff, as in
[STCOM-1436](https://folio-org.atlassian.net/browse/STCOM-1436) and
[STCOM-1447](https://folio-org.atlassian.net/browse/STCOM-1447).

This PR wraps the children of an `<AccordionSet>` in a
`<StripesOverlayWrapper>` which renders all of the overlay components '
menus - `<Dropdown>`, `<Datepicker>`, `<Selection>` etc - to the
`#OverlayContainer`.

Includes tests for the `<StripesOverlayWrapper>` within the
`<AccordionSet>`.
Minor updates to the accordion test suite - linting and usage of
`mountWithContext` vs `mount` to quiet the console about the missing
`Intl` provider.

Additionally, this PR quiets console noise by supplying refs to
`Hotkeys`' `attach` prop instead of the ref'd elements. (In
`<Accordion>` and..well...`<AdvancedSearch>`.