STCOM-1442 itemToView bug. Table scrolls unnecessarily when checkboxes are clicked. (#2472)
Cleanup needed... already improved behavior in the storybook
item-to-view story.

1) previous item capturing was done from `event.target` rather than
'item' aka 'row'. Sometimes `target` might be a `<Checkbox>` that's
aligned middle, so measurements are different from expected. The
'markPosition' logic now uses `.closest((`rowFormatterContainer`)` to
get the closest row internally rendered element, and measure based on
that.

2) previous item capturing happened on focus. While this is fine when a
row is selected, it's awkward because it won't happen again while the
row is focus - and interactions like checkbox checks and stuff can
happen again in the focus state, so the item capture wouldn't be updated
correctly. I've added an `onMouseDown` handler for POC-. ~~although this
leaves keyboard interaction out... will seek to improve.~~

This PR resolves the bug issue when checkboxes or interactive elements
within the row are clicked.