STCOM-1475 AuditLogModal: accept itemFormatter prop (#2519)
# Description
These changes add the possibility to pass a customItemFormatter to the
AuditLogModal and add to this Formatter the collectionName.

# Why is this change necessary?
While adding additionalCallNumbers to the inventorys item history as
requested per this user story:
https://folio-org.atlassian.net/browse/UIIN-3540 we noticed, that
objects are currently being displayed as unordered, unnamed lists. This
seems to be deficient as the user has to guess which value belongs to
which property.

By adding a custom formatter, we are able to display these objects like
this:

<img width="2978" height="1269" alt="Screenshot from 2025-12-11
15-54-34"
src="https://github.com/user-attachments/assets/8ea804fd-6dcf-4dea-8825-c931c1b9b171"
/>

# Related changes

## Mod-Audit
https://github.com/folio-org/mod-audit/pull/229

## UI-Inventory
https://github.com/folio-org/ui-inventory/pull/2950

---------

Co-authored-by: elsenhans <elsenhans@users.noreply.github.com>
Co-authored-by: elsenhans <48911833+elsenhans@users.noreply.github.com>